Revert "Avoid global ordinals in composite aggregation (#74559)"
This reverts commit 5cfcb2f4dbcb52fae24ba44574f8eaf1bcf2299b.

 Conflicts:
	server/src/main/java/org/elasticsearch/search/aggregations/bucket/composite/CompositeValuesCollectorQueue.java
	server/src/main/java/org/elasticsearch/search/aggregations/bucket/composite/OrdinalValuesSource.java
	server/src/main/java/org/elasticsearch/search/aggregations/bucket/composite/TermsValuesSourceBuilder.java
	server/src/test/java/org/elasticsearch/search/aggregations/bucket/composite/SingleDimensionValuesSourceTests.java
