+++ server/src/main/java/org/elasticsearch/search/aggregations/bucket/composite/SingleDimensionValuesSource.java
@@ -160,12 +160,4 @@ abstract class SingleDimensionValuesSource<T extends Comparable<T>> implements R
         }
         return true;
     }
-
-    /**
-     * Whether this values source only guarantees stable hashes for {@link #hashCode(int)} and {@link #hashCodeCurrent()}
-     * in the context of a single LeafReader or whether rehashing is required when switching LeafReaders.
-     */
-    public boolean requiresRehashingWhenSwitchingLeafReaders() {
-        return false;
-    }
 }
