+++ server/src/test/java/org/elasticsearch/search/aggregations/bucket/composite/CompositeValuesCollectorQueueTests.java
@@ -270,9 +270,10 @@ public class CompositeValuesCollectorQueueTests extends AggregatorTestCase {
                     );
                 } else if (types[i].clazz == BytesRef.class) {
                     if (forceMerge) {
-                        sources[i] = new OrdinalValuesSource(
+                        // we don't create global ordinals but we test this mode when the reader has a single segment
+                        // since ordinals are global in this case.
+                        sources[i] = new GlobalOrdinalValuesSource(
                             bigArrays,
-                            (b) -> {},
                             fieldType,
                             context -> DocValues.getSortedSet(context.reader(), fieldType.name()),
                             DocValueFormat.RAW,
